Separeta animation enabling from config enabling in store
